requirements.txt
# Core dependencies for DeepSeek-OCR-2 App
# For CPU-only inference (recommended for most users)

# DeepSeek-OCR-2 dependencies
torch>=2.6.0
transformers>=4.46.3
tokenizers>=0.20.3
PyMuPDF>=1.24.0  # fitz
img2pdf>=0.5.0
einops>=0.8.0
easydict>=1.13
addict>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0

# PageIndex dependencies
openai>=1.101.0
PyPDF2>=3.0.1
python-dotenv>=1.1.0
tiktoken>=0.11.0
pyyaml>=6.0.2

# Web application dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Optional: for GPU support
# torch>=2.6.0 with CUDA (install from pytorch.org based on your CUDA version)
# flash-attn>=2.7.3  # if using flash attention

# Note: DeepSeek-OCR-2 model weights must be downloaded separately
# from https://huggingface.co/deepseek-ai/DeepSeek-OCR-2